requirements.txt
streamlit==1.28.1
openai
Pillow
requests
streamlit-drawable-canvas 
opencv-python 
numpy